Bot/requirements.txt
Flask==3.0.1
PyGithub==2.1.1
Requests==2.31.0
gunicorn==22.0.0
apscheduler==3.10.4
pytz==2024.1